pyproject.toml
[project]
name = "w2-dbquery"
version = "0.1.0"
description = "Natural Language to SQL Query Demo"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "openai>=1.12.0",
    "pydantic>=2.6.0",
]

[project.scripts]
dev = "uvicorn backend.main:app --reload"